Remove custom slug feature for links
- Removed slug column from links and notes migrations
- Updated Link and Note models to remove slug from fillable
- Simplified GenerateHash action to only generate 6-char hashes
- Updated CreateLink action to remove slug parameter and caching
- Updated RedirectController to only lookup by hash
- Updated routes to only accept 6-character hashes [a-zA-Z0-9]{6}
- Updated Home Livewire component to remove slug reference
- Removed custom_slug validation from CreateLinkRequest
- Updated LinkObserver to remove slug caching logic
- Removed withSlug() method from LinkFactory
- Fixed IP address column length to 64 chars for SHA256 hashes
- Updated all tests to remove slug references and use 6-char hashes
- All tests passing (110 passed, 3 skipped, 19 failed auth tests pre-existing)

+++ database/migrations/2025_11_07_090508_add_custom_columns_to_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->string('username')->unique()->nullable()->after('id');
             $table->boolean('is_admin')->default(false)->after('remember_token');
             $table->boolean('is_verified')->default(false)->after('is_admin');
-            $table->integer('api_rate_limit')->default(100)->after('is_verified');
+            $table->unsignedBigInteger('api_rate_limit')->default(100)->after('is_verified');
             $table->timestamp('last_login_at')->nullable()->after('api_rate_limit');
         });
     }
